+++ front/components/actions/tables_query/TablesQueryActionDetails.tsx
@@ -6,15 +6,15 @@ import {
   TableIcon,
   useSendNotification,
 } from "@dust-tt/sparkle";
+import { CodeBlock } from "@dust-tt/sparkle";
+import { ContentBlockWrapper } from "@dust-tt/sparkle";
+import { Markdown } from "@dust-tt/sparkle";
 import type { LightWorkspaceType, TablesQueryActionType } from "@dust-tt/types";
 import { getTablesQueryResultsFileTitle } from "@dust-tt/types";
 import { useCallback } from "react";
 
 import { ActionDetailsWrapper } from "@app/components/actions/ActionDetailsWrapper";
 import type { ActionDetailsComponentBaseProps } from "@app/components/actions/types";
-import { CodeBlock } from "@app/components/assistant/markdown/CodeBlock";
-import { ContentBlockWrapper } from "@app/components/assistant/markdown/ContentBlockWrapper";
-import { RenderMessageMarkdown } from "@app/components/assistant/markdown/RenderMessageMarkdown";
 
 export function TablesQueryActionDetails({
   action,
@@ -79,7 +79,7 @@ function QueryThinking({ action }: { action: TablesQueryActionType }) {
           icon={InformationCircleIcon}
           size="lg"
         >
-          <RenderMessageMarkdown
+          <Markdown
             content={thinking}
             isStreaming={false}
             textSize="sm"
