+++ front/components/actions/websearch/utils.ts
@@ -1,6 +1,6 @@
 import type { WebsearchActionType, WebsearchResultType } from "@dust-tt/types";
 
-import type { MarkdownCitation } from "@app/components/assistant/markdown/MarkdownCitation";
+import type { MarkdownCitation } from "@app/components/markdown/MarkdownCitation";
 
 export function makeWebsearchResultsCitation(
   result: WebsearchResultType
