+++ front/components/assistant/conversation/UserMessage.tsx
@@ -2,11 +2,18 @@ import type {
   ConversationMessageEmojiSelectorProps,
   ConversationMessageSizeType,
 } from "@dust-tt/sparkle";
-import { ConversationMessage } from "@dust-tt/sparkle";
+import { ConversationMessage, Markdown } from "@dust-tt/sparkle";
 import type { UserMessageType, WorkspaceType } from "@dust-tt/types";
+import { useMemo } from "react";
+import type { Components } from "react-markdown";
+import type { PluggableList } from "react-markdown/lib/react-markdown";
 
 import { AgentSuggestion } from "@app/components/assistant/conversation/AgentSuggestion";
-import { RenderMessageMarkdown } from "@app/components/assistant/markdown/RenderMessageMarkdown";
+import { CiteBlock, citeDirective } from "@app/components/markdown/CiteBlock";
+import {
+  MentionBlock,
+  mentionDirective,
+} from "@app/components/markdown/MentionBlock";
 
 interface UserMessageProps {
   citations?: React.ReactElement[];
@@ -27,6 +34,19 @@ export function UserMessage({
   owner,
   size,
 }: UserMessageProps) {
+  const additionalMarkdownComponents: Components = useMemo(
+    () => ({
+      sup: CiteBlock,
+      mention: MentionBlock,
+    }),
+    []
+  );
+
+  const additionalMarkdownPlugins: PluggableList = useMemo(
+    () => [citeDirective(), mentionDirective()],
+    []
+  );
+
   return (
     <ConversationMessage
       pictureUrl={message.user?.image || message.context.profilePictureUrl}
@@ -39,10 +59,12 @@ export function UserMessage({
     >
       <div className="flex flex-col gap-4">
         <div>
-          <RenderMessageMarkdown
+          <Markdown
             content={message.content}
             isStreaming={false}
             isLastMessage={isLastMessage}
+            additionalMarkdownComponents={additionalMarkdownComponents}
+            additionalMarkdownPlugins={additionalMarkdownPlugins}
           />
         </div>
         {message.mentions.length === 0 && isLastMessage && (
