+++ front/components/markdown/MarkdownCitation.tsx
@@ -1,0 +1,8 @@
+import { CitationType } from "@sparkle/components/Citation";
+
+export interface MarkdownCitation {
+  description?: string;
+  href?: string;
+  title: string;
+  type: CitationType;
+}
