+++ front/components/markdown/MentionBlock.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { visit } from "unist-util-visit";
 
 export function MentionBlock({ agentName }: { agentName: string }) {
@@ -9,6 +10,7 @@ export function MentionBlock({ agentName }: { agentName: string }) {
 }
 
 export function mentionDirective() {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   return (tree: any) => {
     visit(tree, ["textDirective"], (node) => {
       if (node.name === "mention" && node.children[0]) {
